Modified the ASE.py
