+++ Assignment/constants.py
@@ -1,0 +1,4 @@
+import string
+
+# Define the character set
+CHAR_SET = string.ascii_letters + string.digits + string.punctuation
